% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact.r
\name{compact}
\alias{compact}
\title{Compact}
\usage{
compact(data, per.sample = TRUE, col = "Height", sim = FALSE, debug = FALSE)
}
\arguments{
\item{data}{data frame with at least 'Sample.Name', 'Marker', 'Allele', and 'Height' columns.}

\item{per.sample}{logical TRUE compact per sample, FALSE for entire dataset (in which case
'Sample.Name' is set to 'Profile' and if sim=TRUE 'Sim' is set to '1' and 'PCR.Vol' is set to the mean).}

\item{col}{character string to specify which column to compact.}

\item{sim}{logical to specify if simulation columns should be replicated (i.e. 'Sim' and 'PCR.Vol').}

\item{debug}{logical for printing debug information.}
}
\value{
data.frame with columns 'Sample.Name', 'Marker', 'Allele', and 'Height'.
}
\description{
Add Identical Peaks.
}
\details{
Add peak heights of identical allele designations together.
}
