% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compactStutter.r
\name{compactStutter}
\alias{compactStutter}
\title{Merge Stutters With Profile.}
\usage{
compactStutter(data, targetcol = NA, stutterin = "PCR.Stutter.",
  debug = FALSE)
}
\arguments{
\item{data}{data.frame to be merged. Columns 'Sample.Name', 'Marker', and
'Allele' are required.}

\item{targetcol}{character name for target column.}

\item{stutterin}{character name for stutter column.}

\item{debug}{logical for printing debug information.}
}
\value{
data.frame with simulation results in columns 'CE.xxx'.
}
\description{
Add peak heights of stutters to the peak heights of alleles.
}
\details{
Merge stutter values to alleles in a simulated dataset.
}

