\name{generateEPGgui}
\alias{generateEPGgui}
\title{Generate EPG GUI}
\usage{
generateEPGgui()
}
\description{
\code{generateEPGgui} is a GUI simplifying the process of
generating an EPG.
}
\details{
A graphical user interface (GUI) simplifying the use of
generating an electropherogram from typing data. Support
different kits (to be improved) by drop-down meny. Manually
enter a profile, or select a data frame from workspace, or
import data from a text file. View and edit the data.
Generate EPG. Possible to save the EPG as a png image file.
}
\examples{
\dontrun{
# Open the graphical user interface.
generateEPGgui()
}
}

