\docType{package}
\name{pcrsim-package}
\alias{pcrsim-package}
\title{Simulation of the forensic DNA process}
\description{
  Forensic DNA process simulator.
}
\details{
  PCRsim is a package for simulating the forensic DNA
  process. \code{pcrsim} opens up a graphical user
  interface which allow the user to enter parameters
  required for the simulation. Once calibrated the program
  can be used to: reduce the laboratory work needed for
  validation of new STR kits, help develop methods for
  interpretaion of DNA evidence, etc. This is a first
  version which is still experimental/under development.

  \tabular{ll}{ Package: \tab pcrsim\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-03-09\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\section{Warning}{
  This package is experimental and has not been thoroughly
  validated.
}
\author{
  Oskar Hansson \email{oskar.hansson@fhi.no}
}
\references{
  Gill, Peter, James Curran, and Keith Elliot. \\u0022 A
  Graphical Simulation Model of the Entire DNA Process
  Associated with the Analysis of Short Tandem Repeat
  Loci\\u0022 Nucleic Acids Research 33, no. 2 (2005):
  632–643. doi:10.1093/nar/gki205.
}
\keyword{package}

