% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmultinomxl.r
\name{rmultinomxl}
\alias{rmultinomxl}
\title{Vectorized Multinomial Distribution For Large Numbers}
\usage{
rmultinomxl(n, size, prob, check = TRUE, debug = FALSE)
}
\arguments{
\item{n}{numeric vector with number of observations.}

\item{size}{numeric vector specifying the number of molecules going into the PCR cycle.}

\item{prob}{numeric non-negative matrix of size (x x K) specifying the probability for the K classes.}

\item{check}{logical for extended checking of parameters.}

\item{debug}{logical to print debug information.}
}
\value{
matrix with simulated results
}
\description{
Simulates one cycle of the PCR process.
}
\details{
Generate multinomially distributed random number vectors using \code{rmultinomial}.
Handles integer overflow by splitting numbers in chunks and call \code{rmultinomial} repeatedly.
}

