\name{pcrsim}
\alias{pcrsim}
\title{GUI for PCR simulator}
\usage{
  pcrsim()
}
\description{
  \code{pcrsim} is a GUI for simulation of the entire DNA
  process.
}
\details{
  This graphical user interface make it very easy to
  simulate the forensic DNA process. Detailes are entered
  in text boxes organised into tabs for the respective
  subprocess. Simulation is performed and the result can be
  viewed within the GUI or plotted as an electropherogram
  (EPG), saved to a text file, or as an R object in the
  global environment. The EPG can be saved as a png-image.
}
\examples{
\dontrun{
# Open the graphical user interface.
pcrsim()
}
}

